% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{A}
\alias{A}
\title{Computes the Adjacency linear operator which maps a vector of weights into
a valid Adjacency matrix.}
\usage{
A(w)
}
\arguments{
\item{w}{weight vector of the graph}
}
\value{
Aw the Adjacency matrix
}
\description{
Computes the Adjacency linear operator which maps a vector of weights into
a valid Adjacency matrix.
}
