% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{altAstar}
\alias{altAstar}
\title{Alternative implementation for the Astar operator.
This is only used for unit testing. Use Astar for
a better performance.}
\usage{
altAstar(M)
}
\arguments{
\item{M}{matrix}
}
\value{
w vector
}
\description{
Alternative implementation for the Astar operator.
This is only used for unit testing. Use Astar for
a better performance.
}
