% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{altLstar}
\alias{altLstar}
\title{Alternative implementation for the Lstar operator.
This is only used for unit testing. Use Lstar for
a better performance.}
\usage{
altLstar(M)
}
\arguments{
\item{M}{matrix}
}
\value{
w vector
}
\description{
Alternative implementation for the Lstar operator.
This is only used for unit testing. Use Lstar for
a better performance.
}
