% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockCoordinateDescent.R
\name{bipartite.psi_update}
\alias{bipartite.psi_update}
\title{Updates the value of psi, the vector whose entries correspond to the
eigenvalues of the adjacency matrix}
\usage{
bipartite.psi_update(V, Aw, lb = -Inf, ub = Inf)
}
\arguments{
\item{V}{matrix whose columns represent the eigen vectors of the Laplacian
matrix in increasing order}

\item{Aw}{Adjacency matrix generated by the weight vector of the graph}
}
\description{
Updates the value of psi, the vector whose entries correspond to the
eigenvalues of the adjacency matrix
}
