% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockCoordinateDescent.R
\name{bipartite.V_update}
\alias{bipartite.V_update}
\title{Updates the value of V, the matrix whose columns represent the eigenvectors
of the Adjacency in increasing order}
\usage{
bipartite.V_update(Aw, n, z)
}
\arguments{
\item{Lw}{Laplacian matrix generated by the weight vector of the graph}

\item{N}{number of nodes}

\item{K}{number of components of the graph}
}
\value{
the updated value of U
}
\description{
Updates the value of V, the matrix whose columns represent the eigenvectors
of the Adjacency in increasing order
}
