% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{blockDiag}
\alias{blockDiag}
\title{Constructs a block diagonal matrix from a list of square matrices}
\usage{
blockDiag(...)
}
\arguments{
\item{...}{list of matrices or individual matrices}
}
\value{
block diagonal matrix
}
\description{
Constructs a block diagonal matrix from a list of square matrices
}
