% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{D}
\alias{D}
\title{Computes the degree operator from the vector of edge weights.}
\usage{
D(w)
}
\arguments{
\item{w}{vector}
}
\value{
Dw vector
}
\description{
Computes the degree operator from the vector of edge weights.
}
