% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Dstar}
\alias{Dstar}
\title{Computes the Dstar operator, i.e., the adjoint of the D operator.}
\usage{
Dstar(w)
}
\arguments{
\item{w}{vector}
}
\value{
Dstar(w) vector
}
\description{
Computes the Dstar operator, i.e., the adjoint of the D operator.
}
