% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fdr}
\alias{fdr}
\title{Computes the false discovery rate between two matrices}
\usage{
fdr(Wtrue, West, eps = 1e-04)
}
\arguments{
\item{Wtrue}{true matrix}

\item{West}{estimated matrix}

\item{eps}{real number such that edges whose values are smaller than eps are
not considered in the computation of the fscore}
}
\description{
Computes the false discovery rate between two matrices
}
\examples{
library(spectralGraphTopology)
X <- L(c(1, 0, 1))
fdr(X, X)
}
