% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fscore}
\alias{fscore}
\title{Computes the fscore between two matrices}
\usage{
fscore(Wtrue, West, eps = 1e-04)
}
\arguments{
\item{Wtrue}{true matrix}

\item{West}{estimated matrix}

\item{eps}{real number such that edges whose values are smaller than eps are
not considered in the computation of the fscore}
}
\description{
Computes the fscore between two matrices
}
\examples{
library(spectralGraphTopology)
X <- L(c(1, 0, 1))
fscore(X, X)
}
