% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{L}
\alias{L}
\title{Computes the Laplacian linear operator which maps a vector of weights into
a valid Laplacian matrix.}
\usage{
L(w)
}
\arguments{
\item{w}{weight vector of the graph}
}
\value{
Lw the Laplacian matrix
}
\description{
Computes the Laplacian linear operator which maps a vector of weights into
a valid Laplacian matrix.
}
