% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockCoordinateDescent.R
\name{lambda_update}
\alias{lambda_update}
\title{Updates the value of lambda, the vector whose entries correspond to the
eigenvalues of the Laplacian matrix}
\usage{
lambda_update(lb, ub, beta, U, w, N, K)
}
\arguments{
\item{lb}{lower bound on the eigenvalues of the Laplacian matrix}

\item{ub}{upper bound on the eigenvalues of the Laplacian matrix}

\item{U}{matrix whose columns represent the eigen vectors of the Laplacian
matrix in increasing order}

\item{w}{weight vector of the graph}

\item{N}{dimension of each data sample}

\item{K}{number of components of the graph}

\item{β}{scalar that controls the strength of the regularization term}
}
\value{
the updated value of λ
}
\description{
Updates the value of lambda, the vector whose entries correspond to the
eigenvalues of the Laplacian matrix
}
