% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockCoordinateDescent.R
\name{laplacian.U_update}
\alias{laplacian.U_update}
\title{Updates the value of U, the matrix whose columns represent the eigen vectors
of the Laplacian in increasing order}
\usage{
laplacian.U_update(Lw, N, K)
}
\arguments{
\item{Lw}{Laplacian matrix generated by the weight vector of the graph}

\item{N}{dimension of each data sample}

\item{K}{number of components of the graph}
}
\value{
the updated value of U
}
\description{
Updates the value of U, the matrix whose columns represent the eigen vectors
of the Laplacian in increasing order
}
