% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockCoordinateDescent.R
\name{laplacian.w_update}
\alias{laplacian.w_update}
\title{Updates the value of w, the weight vector of the graph Laplacian}
\usage{
laplacian.w_update(w, Lw, U, beta, lambda, n, Kmat)
}
\arguments{
\item{w}{weight vector of the graph}

\item{Lw}{Laplacian matrix generated by the weight vector of the graph}

\item{U}{matrix whose columns represent the eigen vectors of the Laplacian
matrix in increasing order}

\item{Kmat}{matrix (see section 1.1 for its definition)}

\item{β}{scalar that controls the strength of the regularization term}

\item{λ}{vector whose entries are the eigenvalues of the Laplacian in
increasing order}

\item{N}{number of nodes}
}
\value{
the updated value of w
}
\description{
Updates the value of w, the weight vector of the graph Laplacian
}
