% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Linv}
\alias{Linv}
\title{Computes the inverse of the L operator.}
\usage{
Linv(M)
}
\arguments{
\item{M}{Laplacian matrix}
}
\value{
w the weight vector of the graph
}
\description{
Computes the inverse of the L operator.
}
