% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Mmat}
\alias{Mmat}
\title{Computes the matrix form of the composition of the operators Lstar and
L, i.e., Lstar o L.}
\usage{
Mmat(n)
}
\arguments{
\item{n}{number of columns/rows}
}
\value{
M the composition of Lstar and L
}
\description{
Computes the matrix form of the composition of the operators Lstar and
L, i.e., Lstar o L.
}
