% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{npv}
\alias{npv}
\title{Computes the negative predictive value between two matrices}
\usage{
npv(Wtrue, West, eps = 1e-04)
}
\arguments{
\item{Wtrue}{true matrix}

\item{West}{estimated matrix}

\item{eps}{real number such that edges whose values are smaller than eps are
not considered in the computation of the fscore}
}
\description{
Computes the negative predictive value between two matrices
}
\examples{
library(spectralGraphTopology)
X <- L(c(1, 0, 1))
npv(X, X)
}
