% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{performance_per_iteration}
\alias{performance_per_iteration}
\title{Given a reference Laplacian matrix, this function computes the relative error,
the negative loglikelihood, and the objective function per iteration for the
spectralGraphTopology algorithm}
\usage{
performance_per_iteration(Lw, T, K, ...)
}
\arguments{
\item{Lw}{reference Laplacian matrix from which random samples will be drawn}

\item{T}{number of samples to be drawn which will be used to estimate L}

\item{K}{number of components of the graph}

\item{...}{extra arguments to be passed to learnGraphTopology}
}
\value{
rel_err sequence of relative error measurements for each iteration

obj_fun objective function value for each iteration

loglike negative loglikelihood value for each iteration
}
\description{
Given a reference Laplacian matrix, this function computes the relative error,
the negative loglikelihood, and the objective function per iteration for the
spectralGraphTopology algorithm
}
