% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Pmat}
\alias{Pmat}
\title{Computes the matrix form of the composition of the operators Astar and
A, i.e., Astar o A.}
\usage{
Pmat(n)
}
\arguments{
\item{n}{number of columns/rows}
}
\value{
M the composition of Astar and A
}
\description{
Computes the matrix form of the composition of the operators Astar and
A, i.e., Astar o A.
}
