% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prial}
\alias{prial}
\title{Compute the prial value between two matrices}
\usage{
prial(Ltrue, Lest, Lscm)
}
\arguments{
\item{Ltrue}{true Laplacian matrix}

\item{Lest}{estimated Laplacian matrix}

\item{Lscm}{estimated Laplacian matrix via the generalized inverse of the
of the sample covariance matrix}
}
\description{
Compute the prial value between two matrices
}
