% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{relative_error}
\alias{relative_error}
\title{Computes the relative error between the true and estimated matrices}
\usage{
relative_error(West, Wtrue)
}
\arguments{
\item{West}{estimated matrix}

\item{Wtrue}{true matrix}
}
\description{
Computes the relative error between the true and estimated matrices
}
\examples{
library(spectralGraphTopology)
X <- L(c(1, 0, 1))
relative_error(X, X)
}
