% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{relativeError}
\alias{relativeError}
\title{Compute the relative error between two matrices}
\usage{
relativeError(Ltrue, Lest)
}
\arguments{
\item{Ltrue}{true Laplacian matrix}

\item{Lest}{estimated Laplacian matrix}
}
\description{
Compute the relative error between two matrices
}
