% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockCoordinateDescent.R
\name{U_update}
\alias{U_update}
\title{Updates the value of U, the matrix whose columns represent the eigen vectors
of the Laplacian in increasing order}
\usage{
U_update(w, N)
}
\arguments{
\item{w}{weight vector of the graph}

\item{N}{dimension of each data sample}
}
\value{
the updated value of U
}
\description{
Updates the value of U, the matrix whose columns represent the eigen vectors
of the Laplacian in increasing order
}
