% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vec}
\alias{vec}
\title{Computes the canonical vec operator, i.e., converts a given n x m matrix M,
into a  nm x 1 column vector vec(M) by stacking the columns of M on top of
one another.}
\usage{
vec(M)
}
\arguments{
\item{M}{input matrix}
}
\value{
w vector such that w = vec(M)
}
\description{
Computes the canonical vec operator, i.e., converts a given n x m matrix M,
into a  nm x 1 column vector vec(M) by stacking the columns of M on top of
one another.
}
