% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vecLmat}
\alias{vecLmat}
\title{Computes the matrix that represents the composition of
the vec and the L operators.}
\usage{
vecLmat(n)
}
\arguments{
\item{n}{the dimension of L}
}
\value{
R matrix such that vec(L(w)) = Rw
}
\description{
Computes the matrix that represents the composition of
the vec and the L operators.
}
