% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockCoordinateDescent.R
\name{w_init}
\alias{w_init}
\title{helper function used to compute initial guesses
for the vector of weights of the Laplacian matrix}
\usage{
w_init(w0, Sinv)
}
\arguments{
\item{w0}{string represeting which initial point to use.
options are "qp" and "naive".}

\item{Sinv}{inverse of the sample covariance matrix.}
}
\description{
helper function used to compute initial guesses
for the vector of weights of the Laplacian matrix
}
