% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockCoordinateDescent.R
\name{w_update}
\alias{w_update}
\title{Updates the value of w, the weight vector of the graph}
\usage{
w_update(w, U, beta, lambda, N, Kmat)
}
\arguments{
\item{w}{weight vector of the graph}

\item{U}{matrix whose columns represent the eigen vectors of the Laplacian
matrix in increasing order}

\item{N}{dimension of each data sample}

\item{Kmat}{matrix (see section 1.1 for its definition)}

\item{β}{scalar that controls the strength of the regularization term}

\item{λ}{vector whose entries are the eigenvalues of the Laplacian in
increasing order}
}
\value{
the updated value of w
}
\description{
Updates the value of w, the weight vector of the graph
}
