% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockCoordinateDescent.R
\name{adjacency.psi_update}
\alias{adjacency.psi_update}
\title{Updates the value of psi, the vector whose entries correspond to the
eigenvalues of the adjacency matrix}
\usage{
adjacency.psi_update(lb, ub, V, Aw)
}
\arguments{
\item{lb}{lower bound on the eigenvalues of the Adjacency matrix}

\item{ub}{upper bound on the eigenvalues of the Adjacency matrix}

\item{V}{matrix whose columns represent the eigen vectors of the Laplacian
matrix in increasing order}

\item{Aw}{Adjacency matrix generated by the weight vector of the graph}

\item{β}{scalar that controls the strength of the regularization term}
}
\description{
Updates the value of psi, the vector whose entries correspond to the
eigenvalues of the adjacency matrix
}
