% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrLaplacianRank.R
\name{cluster_k_component_graph}
\alias{cluster_k_component_graph}
\title{Cluster a k-component graph from data using the Constrained Laplacian Rank algorithm

Cluster a k-component graph on the basis of an observed data matrix}
\usage{
cluster_k_component_graph(Y, k = 1, m = 5, lmd = 1,
  eig_tol = 1e-09, edge_tol = 1e-06, maxiter = 1000)
}
\arguments{
\item{Y}{a pxn data matrix, where p is the number of nodes and n is the number of
features (or data points per node)}

\item{k}{the number of components of the graph}

\item{m}{the maximum number of possible connections for a given node used
to build an affinity matrix}

\item{lmd}{L2-norm regularization hyperparameter}

\item{eig_tol}{value below which eigenvalues are considered to be zero}

\item{edge_tol}{value below which edge weights are considered to be zero}

\item{maxiter}{the maximum number of iterations}
}
\value{
A list containing the following elements:
\item{\code{Laplacian}}{the estimated Laplacian Matrix}
\item{\code{Adjacency}}{the estimated Adjacency Matrix}
\item{\code{eigvals}}{the eigenvalues of the Laplacian Matrix}
\item{\code{lmd_seq}}{sequence of lmd values at every iteration}
\item{\code{elapsed_time}}{elapsed time at every iteration}
}
\description{
Cluster a k-component graph from data using the Constrained Laplacian Rank algorithm

Cluster a k-component graph on the basis of an observed data matrix
}
\examples{
library(spectralGraphTopology)
library(igraph)
library(pals)
library(kernlab)
data(iris)
n <- nrow(iris)
Y <- t(matrix(as.numeric(unlist(iris[, 1:4])), nrow = nrow(iris)))
names <- c(as.character(iris[, 5]))
unique_names <- c(as.character(unique(iris[, 5])))
graph <- cluster_k_component_graph(t(Y), k = 3, m = 5)
net <- graph_from_adjacency_matrix(graph$Adjacency, mode = "undirected", weighted = TRUE)
clusters <- array(0, length(names))
for (i in c(1:length(unique_names)))
  clusters[unique_names[i] == names] <- i
V(net)$cluster <- clusters
colors <- c("#34495E", "#706FD3", "#FF5252")
E(net)$color <- apply(as.data.frame(get.edgelist(net)), 1,
                     function(x) ifelse(V(net)$cluster[x[1]] == V(net)$cluster[x[2]],
                                        colors[V(net)$cluster[x[1]]], brewer.greys(5)[2]))
V(net)$color <- colors[clusters]
gr = .5 * (1 + sqrt(5))
setEPS()
postscript("iris-clr.ps", family = "Times", height = 5, width = gr * 3.5)
plot(net, vertex.label = NA, vertex.size = 3)
dev.off()
}
\references{
Nie, Feiping and Wang, Xiaoqian and Jordan, Michael I. and Huang, Heng.
            The Constrained Laplacian Rank Algorithm for Graph-based Clustering, 2016,
            AAAI'16. http://dl.acm.org/citation.cfm?id=3016100.3016174
}
\author{
Ze Vinicius and Daniel Palomar
}
