% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{relative_error}
\alias{relative_error}
\title{Computes the relative error between two matrices}
\usage{
relative_error(A, B)
}
\arguments{
\item{A}{first matrix}

\item{B}{second matrix}
}
\description{
Computes the relative error between two matrices
}
\examples{
library(spectralGraphTopology)
X <- L(c(1, 0, 1))
relative_error(X, X)
}
